rename snack-bar service
