+++ angular/src/app/app.component.spec.ts
@@ -6,7 +6,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { SidenavModule } from './sidenav/sidenav.module';
 import { CoreModule } from './core/core.module';
 
-import { SnackBarService } from './core/snack/snack.service';
+import { SnackBarService } from './core/snack-bar/snack-bar.service';
 import { WindowService } from './core/window/window.service';
 import { UserAreaService } from './user-area/shared/user-area.service';
 import { AuthService } from './core/authentication/auth.service';
