+++ angular/src/app/book-table/book-table.component.ts
@@ -4,7 +4,7 @@ import { MatDialog, MatCheckbox } from '@angular/material';
 import { BookTableDialogComponent } from './book-table-dialog/book-table-dialog.component';
 import { InvitationDialogComponent } from './invitation-dialog/invitation-dialog.component';
 import { WindowService } from '../core/window/window.service';
-import { SnackBarService } from '../core/snack/snack.service';
+import { SnackBarService } from '../core/snack-bar/snack-bar.service';
 import { emailValidator } from '../shared/directives/email-validator.directive';
 import { last } from 'lodash';
 import { BookingInfo } from 'app/shared/backend-models/interfaces';
