+++ angular/src/app/core/core.module.ts
@@ -53,7 +53,7 @@ import {
 import { MomentModule } from 'ngx-moment';
 
 import { WindowService } from './window/window.service';
-import { SnackBarService } from './snack/snack.service';
+import { SnackBarService } from './snack-bar/snack-bar.service';
 import { HttpRequestInterceptorService } from './interceptor/http-request-interceptor.service';
 import { AuthGuardService } from './authentication/auth-guard.service';
 import { AuthService } from './authentication/auth.service';
