+++ angular/src/app/email-confirmations/email-confirmations.component.ts
@@ -1,6 +1,6 @@
 import { InvitationResponse } from '../shared/view-models/interfaces';
 import { EmailConfirmationsService } from './shared/email-confirmations.service';
-import { SnackBarService } from '../core/snack/snack.service';
+import { SnackBarService } from '../core/snack-bar/snack-bar.service';
 import { ActivatedRoute, Router, ParamMap } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
