+++ angular/src/app/menu/menu-card/menu-card.component.spec.ts
@@ -3,7 +3,7 @@ import { CoreModule } from '../../core/core.module';
 
 import { SidenavService } from '../../sidenav/shared/sidenav.service';
 import { MenuService } from '../shared/menu.service';
-import { SnackBarService } from '../../core/snack/snack.service';
+import { SnackBarService } from '../../core/snack-bar/snack-bar.service';
 import { AuthService } from '../../core/authentication/auth.service';
 
 import { MenuCardComponent } from './menu-card.component';
