+++ angular/src/app/menu/menu.component.spec.ts
@@ -8,7 +8,7 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { SidenavService } from '../sidenav/shared/sidenav.service';
 import { MenuService } from './shared/menu.service';
-import { SnackBarService } from '../core/snack/snack.service';
+import { SnackBarService } from '../core/snack-bar/snack-bar.service';
 import { AuthService } from '../core/authentication/auth.service';
 import { MatSlider, MatInput } from '@angular/material';
