+++ angular/src/app/sidenav/sidenav.component.spec.ts
@@ -4,7 +4,7 @@ import { RouterTestingModule } from '@angular/router/testing';
 
 import { SidenavService } from './shared/sidenav.service';
 import { PriceCalculatorService } from './shared/price-calculator.service';
-import { SnackBarService } from '../core/snack/snack.service';
+import { SnackBarService } from '../core/snack-bar/snack-bar.service';
 
 import { SidenavOrderComponent } from './sidenav-order/sidenav-order.component';
 import { SidenavComponent } from './sidenav.component';
